JERKY: consolidation
