+++ protocol-jerky/src/main/scala/org/ensime/jerky/JerkyFormats.scala
@@ -1,6 +1,6 @@
 // Copyright: 2010 - 2016 https://github.com/ensime/ensime-server/graphs
 // License: http://www.gnu.org/licenses/gpl-3.0.en.html
-package org.ensime.jerk
+package org.ensime.jerky
 
 import spray.json._
 import fommil.sjs._
@@ -10,7 +10,7 @@ import org.ensime.api._
 
 import org.ensime.util.file._
 
-private object JerkConversions extends DefaultJsonProtocol with FamilyFormats {
+private object JerkyConversions extends DefaultJsonProtocol with FamilyFormats {
   // This part of the code is brought to you by the words "accidental"
   // and "complexity".
   //
@@ -42,24 +42,14 @@ private object JerkConversions extends DefaultJsonProtocol with FamilyFormats {
   // some of the case classes use the keyword `type`, so we need a better default
   override implicit def coproductHint[T: Typeable]: CoproductHint[T] = new FlatCoproductHint[T]("typehint")
 
-  val RpcRequestFormat: RootJsonFormat[RpcRequest] = cachedImplicit
-  val EnsimeServerMessageFormat: RootJsonFormat[EnsimeServerMessage] = cachedImplicit
+  implicit val RpcRequestEnvelopeFormat: RootJsonFormat[RpcRequestEnvelope] = cachedImplicit
+  implicit val RpcResponseEnvelopeFormat: RootJsonFormat[RpcResponseEnvelope] = cachedImplicit
 
 }
 
-object JerkFormats {
-  implicit val RpcRequestFormat: RootJsonFormat[RpcRequest] =
-    JerkConversions.RpcRequestFormat
-  implicit val EnsimeServerMessageFormat: RootJsonFormat[EnsimeServerMessage] =
-    JerkConversions.EnsimeServerMessageFormat
-}
-
-object JerkEnvelopeFormats {
-  import JerkFormats._
-  import DefaultJsonProtocol._
-
-  // for the transitional SWANK (avoid derivations)
-  implicit val RpcRequestEnvelopeFormat: RootJsonFormat[RpcRequestEnvelope] = jsonFormat2(RpcRequestEnvelope)
-  implicit val RpcResponseEnvelopeFormat: RootJsonFormat[RpcResponseEnvelope] = jsonFormat2(RpcResponseEnvelope)
-
+object JerkyFormats {
+  implicit val RpcRequestEnvelopeFormat: RootJsonFormat[RpcRequestEnvelope] =
+    JerkyConversions.RpcRequestEnvelopeFormat
+  implicit val RpcResponseEnvelopeFormat: RootJsonFormat[RpcResponseEnvelope] =
+    JerkyConversions.RpcResponseEnvelopeFormat
 }
