+++ protocol-jerky/src/test/scala/org/ensime/jerky/SprayJsonTestSupport.scala
@@ -1,6 +1,6 @@
 // Copyright: 2010 - 2016 https://github.com/ensime/ensime-server/graphs
 // License: http://www.gnu.org/licenses/gpl-3.0.en.html
-package org.ensime.jerk
+package org.ensime.jerky
 
 import org.scalatest._
 import spray.json._
