+++ server/src/main/scala/org/ensime/server/Server.scala
@@ -126,7 +126,6 @@ object Server {
 
     val protocol: Protocol = propOrElse("ensime.protocol", "swank") match {
       case "swank" => new SwankProtocol
-      case "jerk" => new JerkProtocol
       case other => throw new IllegalArgumentException(s"$other is not a valid ENSIME protocol")
     }
